refactor roles
