+++ packages/lib/src/index.ts
@@ -28,6 +28,7 @@ export * from "./phone-number";
 export * from "./plandetransport";
 export * from "./region-and-departments";
 export * from "./roles";
+export * from "./rolesConfig";
 export * from "./routes";
 export * from "./sessions";
 export * from "./SNUpport";
@@ -41,4 +42,4 @@ export * from "./utils/parseEnv";
 export * from "./utils/request";
 export * from "./utils/csvGenerator";
 export * from "./young";
-export * from "./permissions";
+export * from "./actions";
