--- packages/lib/src/permissions/transport/transportActions.ts
@@ -1,9 +1,0 @@
-export const transportActions = {
-  UPDATE: "TRANSPORT:UPDATE",
-  UPDATE_PDR_ID: "TRANSPORT:UPDATE_PDR_ID",
-  UPDATE_PDR_SCHEDULE: "TRANSPORT:UPDATE_PDR_SCHEDULE",
-  UPDATE_TYPE: "TRANSPORT:UPDATE_TYPE",
-  UPDATE_SESSION_ID: "TRANSPORT:UPDATE_SESSION_ID",
-  UPDATE_CENTER_SCHEDULE: "TRANSPORT:UPDATE_CENTER_SCHEDULE",
-  NOTIFY_AFTER_UPDATE: "TRANSPORT:NOTIFY_AFTER_UPDATE",
-};
